+++ week13/BOJ_13458_시험_감독/이나연.java
@@ -1,0 +1,30 @@
+import java.io.*;
+import java.util.StringTokenizer;
+
+public class Main {
+
+    public static void main(String[] args) throws IOException {
+        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
+        int N = Integer.parseInt(br.readLine());
+        StringTokenizer st = new StringTokenizer(br.readLine());
+        int[] A = new int[N];
+        for (int i = 0; i < N; i++) {
+            A[i] = Integer.parseInt(st.nextToken());
+        }
+        st = new StringTokenizer(br.readLine());
+        int B = Integer.parseInt(st.nextToken());
+        int C = Integer.parseInt(st.nextToken());
+
+        long count = 0;
+        for (int i = 0; i < N; i++) {
+            A[i] -= B;
+            count++;
+            if (A[i] > 0) {
+                count += Math.ceil((double) A[i] / C);
+            }
+        }
+
+        System.out.println(count);
+    }
+
+}
